requirements.txt
google-generativeai
chardet
